Add Monza Waypoints + Add Trace Waypoints Script

+++ roar_py_carla/roar_py_carla/utils/convert_coordinate.py
@@ -8,8 +8,7 @@ def location_from_carla(location : carla.Location) -> np.ndarray:
     return np.array([location.x, -location.y, location.z], dtype=np.float32)
 
 def location_to_carla(location : np.ndarray) -> carla.Location:
-    location = location.astype(np.float32)
-    return carla.Location(x=location[0], y=-location[1], z=location[2])
+    return carla.Location(x=float(location[0]), y=float(-location[1]), z=float(location[2]))
 
 def rotation_from_carla(rotation : carla.Rotation) -> np.ndarray:
     rot_deg = np.array([
@@ -19,7 +18,7 @@ def rotation_from_carla(rotation : carla.Rotation) -> np.ndarray:
 
 def rotation_to_carla(rotation : np.ndarray) -> carla.Rotation:
     rot_deg = np.rad2deg(rotation).astype(np.float32)
-    return carla.Rotation(roll=rot_deg[0], pitch=-rot_deg[1], yaw=-rot_deg[2])
+    return carla.Rotation(roll=float(rot_deg[0]), pitch=-float(rot_deg[1]), yaw=-float(rot_deg[2]))
 
 def transform_from_carla(transform : carla.Transform) -> typing.Tuple[np.ndarray, np.ndarray]:
     location = location_from_carla(transform.location)
